動作確認のコードとデータ取り込みのコードとでファイルを分割
